Isolate workspace snapshot reconciliation failures (#2265)
